Bug fixes and starting tests for cw721-roles

+++ contracts/external/cw721-roles/src/msg.rs
@@ -3,7 +3,7 @@ use cosmwasm_std::CustomMsg;
 
 #[cw_serde]
 pub struct MetadataExt {
-    pub weight: u32,
+    pub weight: u64,
 }
 
 #[cw_serde]
